Abstract out a BmcDeformableMirror parent class
