+++ catkit2/services/bmc_deformable_mirror_hardware/bmc_deformable_mirror_hardware.py
@@ -1,10 +1,8 @@
-from catkit2.base_services.deformable_mirror import DeformableMirrorService
+from catkit2.base_services.bmc_deformable_mirror import BmcDeformableMirror
 
 import sys
 import os
-import threading
 import numpy as np
-from astropy.io import fits
 
 try:
     sdk_path = os.environ.get('CATKIT_BOSTON_SDK_PATH')
@@ -17,44 +15,14 @@ except ImportError:
     raise
 
 
-class BmcDeformableMirrorHardware(DeformableMirrorService):
+class BmcDeformableMirrorHardware(BmcDeformableMirror):
     def __init__(self):
         super().__init__('bmc_deformable_mirror_hardware')
 
-        self.serial_number = self.config['serial_number']
-        self.flat_map_fname = self.config['flat_map_fname']
-        self.gain_map_fname = self.config['gain_map_fname']
-        self.max_volts = self.config['max_volts']
-        self.dac_bit_depth = self.config['dac_bit_depth']
-        self._discretized_voltages = None
-        self._discretized_surface = None
-
         self.device_command_index = self.config.get('device_command_index', 0)
-
-        self.lock = threading.Lock()
 
     def open(self):
         super().open()
-
-        self.flat_map = fits.getdata(self.flat_map_fname)
-        if self.flat_map.ndim <= 1:
-            raise ValueError(f'The provided flat map for {self.service_id} needs to be at least a 2D array.')
-        elif self.flat_map.ndim == 2:
-            self.flat_map = np.expand_dims(self.flat_map, axis=0)
-        # Convert to DM command
-        self.flat_map_command = self.flat_map[self.device_actuator_mask]
-
-        self.gain_map = fits.getdata(self.gain_map_fname)
-        if self.gain_map.ndim <= 1:
-            raise ValueError(f'The provided gain map for {self.service_id} needs to be at least a 2D array.')
-        elif self.gain_map.ndim == 2:
-            self.gain_map = np.expand_dims(self.gain_map, axis=0)
-        # Convert to DM command
-        self.gain_map_command = self.gain_map[self.device_actuator_mask]
-
-        with np.errstate(divide='ignore', invalid='ignore'):
-            self.gain_map_inv_command = 1 / self.gain_map_command    # TODO: Is this still correct?
-            self.gain_map_inv_command[np.abs(self.gain_map_command) < 1e-10] = 0
 
         self.device = bmc.BmcDm()
         status = self.device.open_dm(self.serial_number)
@@ -78,13 +46,7 @@ class BmcDeformableMirrorHardware(DeformableMirrorService):
         super().close()
 
     def send_surface(self, total_surface):
-        # Compute the voltages from the requested total surface.
-        voltages = self.flat_map_command + total_surface * self.gain_map_inv_command
-        voltages /= self.max_volts
-        voltages = np.clip(voltages, 0, 1)
-
-        self.discretized_voltages = voltages
-        self.discretized_surface = total_surface
+        super().send_surface(total_surface)
 
         # Convert to hardware command format
         device_command = self.dm_command_to_device_command(self.discretized_voltages)
@@ -95,32 +57,6 @@ class BmcDeformableMirrorHardware(DeformableMirrorService):
 
             if status != bmc.NO_ERR:
                 raise RuntimeError(f'Failed to send data: {self.device.error_string(status)}.')
-
-        # Submit discretized surface and voltages to data streams.
-        self.total_surface.submit_data(self.discretized_surface)
-        self.total_voltage.submit_data(self.discretized_voltages)
-
-    @property
-    def discretized_voltages(self):
-        return self._discretized_voltages
-
-    @discretized_voltages.setter
-    def discretized_voltages(self, voltages):
-        values = voltages
-        if self.dac_bit_depth is not None:
-            values = (np.floor(voltages * (2**self.dac_bit_depth))) / (2**self.dac_bit_depth)
-        self._discretized_voltages = values
-
-    @property
-    def discretized_surface(self):
-        return self._discretized_surface
-
-    @discretized_surface.setter
-    def discretized_surface(self, total_surface):
-        values = total_surface
-        if self.dac_bit_depth is not None:
-            values = (self.discretized_voltages * self.max_volts - self.flat_map_command) * self.gain_map_command
-        self._discretized_surface = values
 
     def dm_command_to_device_command(self, dm_command):
         device_command = np.zeros(self.device_command_length)
